Add option to load libclang from environment variable
This update introduces a check for the LIBCLANG_PATH environment variable at the beginning of the static block. If the variable is defined, the system will load libclang from the provided path, otherwise, it will default to loading libclang from the temporary directory. This allows developers to override the default libclang path when necessary.

+++ klang/jextract/src/main/java/org/openjdk/jextract/clang/libclang/RuntimeHelper.java
@@ -53,18 +53,22 @@ final class RuntimeHelper {
             (size, align) -> Arena.ofAuto().allocate(size, align);
 
     static {
-        var libraryFile = new File(getTemporaryDirectory() + inferLibraryFileName());
-        if (!libraryFile.exists()) {
-            var embeddedLibraryFile = findFileInClasspath(inferEmbeddedLibraryFileName());
-            try {
-                copyInputStreamToFile(embeddedLibraryFile, libraryFile.getAbsolutePath());
-            } catch (IOException e) {
-                throw new RuntimeException(e);
+        if (System.getenv("LIBCLANG_PATH") != null) {
+            System.load(System.getenv("LIBCLANG_PATH"));
+        } else {
+            var libraryFile = new File(getTemporaryDirectory() + inferLibraryFileName());
+            if (!libraryFile.exists()) {
+                var embeddedLibraryFile = findFileInClasspath(inferEmbeddedLibraryFileName());
+                try {
+                    copyInputStreamToFile(embeddedLibraryFile, libraryFile.getAbsolutePath());
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
             }
+            libraryFile.deleteOnExit();
+            System.load(libraryFile.getAbsolutePath());
         }
-        libraryFile.deleteOnExit();
 
-        System.load(libraryFile.getAbsolutePath());
 
         SymbolLookup loaderLookup = SymbolLookup.loaderLookup();
         SYMBOL_LOOKUP = name -> loaderLookup.find(name).or(() -> LINKER.defaultLookup().find(name));
